# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This service provide functionality for storing and managing privat telegram channels with subscription based payments for acessing content.",
        "title": "Telegram Subs API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/account/list": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "get account list. Only administrator can get list of accounts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "List Accounts",
                "parameters": [
                    {
                        "description": "account list request filter",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.FindAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Account"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/account/{acc_id}": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "get account by id. Creator can get only his account. Administrator can get any account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Get Account",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "account id",
                        "name": "acc_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Account"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "deletes an account. Only administrator can delete accounts.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Delete Account",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "account id",
                        "name": "acc_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/account/{acc_id}/email": {
            "patch": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "updates email for current user. Admin can update email for any user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Update Email",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "account id",
                        "name": "acc_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "account new email",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.updateAccountEmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/account/{acc_id}/password": {
            "patch": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "updates password for current account. Admin can change password without provieding an old password. Admin can update password for any user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Update Password",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "account id",
                        "name": "acc_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "account old and new password",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.updateAccountPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/account/{acc_id}/user": {
            "patch": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "updates telegram user conected to this account. Admin, can update email for any user. Either of one of the fields must be provided",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "Update TgUser",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "account id",
                        "name": "acc_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "account new email",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.updateAccountUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.Account": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "integer"
                },
                "chan_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "role": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "domain.FindAccountRequest": {
            "type": "object",
            "properties": {
                "page_settings": {
                    "description": "Username *struct {\n\tLike string `json:\"LIKE\"`\n\tEq   string `json:\"EQ\"`\n} `json:\"username\"`\nEmail *struct {\n\tLike string `json:\"LIKE\"`\n\tEq   string `json:\"EQ\"`\n} `json:\"email\"`",
                    "type": "object",
                    "properties": {
                        "page_number": {
                            "type": "integer"
                        },
                        "page_size": {
                            "type": "integer"
                        }
                    }
                }
            }
        },
...
        "http.updateAccountEmailRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "http.updateAccountPasswordRequest": {
            "type": "object",
            "properties": {
                "new_password": {
                    "type": "string"
                },
                "old_password": {
                    "type": "string"
                }
            }
        },
        "http.updateAccountUserRequest": {
            "type": "object",
            "properties": {
                "telegram_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "authorization",
            "in": "header"
        }
    }
}